One of the blobs is red
The blob indicating the parent Major scale is now red.
The pie chart starts off showing the C Major scale (C Ionian) and then
each degree in turn is flattened until the C degree itself is flattened
to B. Then each degree of the new B Major scale is flattened, and so on.