I modified wantpausebehaviour
Instead of being a boolean, _pause is now an int which counts down from
an initial value to zero.
The idea here is that when _pause reaches zero a blob is animated. But
before _pause reaches zero no blobs move, which makes it easier to
visually read the current Mode degrees from the blob positions.