Clarification
Originally I was adding the delta angle *before* placing the blob, so
blob[0] was at 2 o'clock instead of at midday, blob[1] was at 4 o'clock,
and so on. This made it rather involved trying to work out the integer
selector for a visual blob from its appearance in the chart. I fixed
this by adding the delta angle after placing the blob, which means
blob[0] is now at midday.